Update 03-data.Rmd
You said a piece of Data was numeric, but for me and visible in your later shows examples it shows as an integer. 
I'm very new to R and english is my second language, so I may simply misunderstand some kind of context here. In anyway, thanks for your project, it's amazing!